+++ exynos4/interfaces/mapper/Gralloc0Mapper.h
@@ -34,10 +34,10 @@ class Gralloc0Mapper : public GrallocMapper {
    private:
     Error registerBuffer(buffer_handle_t bufferHandle) override;
     void unregisterBuffer(buffer_handle_t bufferHandle) override;
-    Error lockBuffer(buffer_handle_t bufferHandle, uint64_t cpuUsage,
+    Error lockBuffer(buffer_handle_t bufferHandle, uint32_t cpuUsage,
                      const IMapper::Rect& accessRegion, int fenceFd,
                      void** outData) override;
-    Error lockBuffer(buffer_handle_t bufferHandle, uint64_t cpuUsage,
+    Error lockBuffer(buffer_handle_t bufferHandle, uint32_t cpuUsage,
                      const IMapper::Rect& accessRegion, int fenceFd,
                      YCbCrLayout* outLayout) override;
     Error unlockBuffer(buffer_handle_t bufferHandle, int* outFenceFd) override;
