+++ exynos4/interfaces/mapper/Gralloc1Mapper.h
@@ -46,10 +46,10 @@ class Gralloc1Mapper : public GrallocMapper {
 
     Error registerBuffer(buffer_handle_t bufferHandle) override;
     void unregisterBuffer(buffer_handle_t bufferHandle) override;
-    Error lockBuffer(buffer_handle_t bufferHandle, uint64_t cpuUsage,
+    Error lockBuffer(buffer_handle_t bufferHandle, uint32_t cpuUsage,
                      const IMapper::Rect& accessRegion, int fenceFd,
                      void** outData) override;
-    Error lockBuffer(buffer_handle_t bufferHandle, uint64_t cpuUsage,
+    Error lockBuffer(buffer_handle_t bufferHandle, uint32_t cpuUsage,
                      const IMapper::Rect& accessRegion, int fenceFd,
                      YCbCrLayout* outLayout) override;
     Error unlockBuffer(buffer_handle_t bufferHandle, int* outFenceFd) override;
